feat(#2148): Move Online class under OnlineCondition class
